Disable admin user bar.
The Wagtail admin user bar is rendered in an iframe, which is blocked by
our Cloud Foundry nginx configuration (via the `x-frame-options`
header). As a result, all scripts loaded after the iframe is inserted
fail for logged-in admin users. As a temporary fix, this patch disables
the admin user bar. In the future @dlapiduz may update the nginx
configuration to let us override `x-frame-options`, in which case we can
revert this change.