Update invoke and related tasks
This changeset updates invoke to the latest stable release and updates our corresponding tasks to work correctly with it.  It will also provide new tasks for retrieving and pushing database dumps to an S3 bucket for the CMS app used by the FEC project.  These tasks will make use of the `cg-migrate-db` plugin for Cloud Foundry written by @jcscottiii.

Huge h/t and thank you to @jcscottiii for all of the hard work and help with the plugin!