Improve CMS README, documentation, and instructions
This changeset encompasses all of the documentation updates we need to make to account for the improvements we are making to the development workflow with the CMS.  It will also bring parity with the other repos that have had their documentation and READMEs updated (e.g., the API in openFEC).

Updates will take place in tandem with the other work happening in related PRs.