Switch local DB to PostgreSQL
This changeset switches our local development setup to make use of PostgreSQL instead of SQLite.  The impetus behind this change is to bring parity across our projects (the API also runs on PostgreSQL) and the fact that the CMS as deployed also runs on PostgreSQL.  We want to take advantage of being able to pull down snapshots of production data to help with future development.  By having the same database backend locally we will be able to accomplish that.