Update README.md for easier install
Went through the installation again and noticed that some users might prefer to run createdb via their default user rather than as the postgres user, so instructions were added for altering the role for the default user to be able to create database. Also noticed that migrate needs to run first in order to have the auth model to createsuperuser.