Add support for cloud.gov GovCloud migration
This changeset includes the necessary changes required to migrate the FEC CMS app to the GovCloud cloud.gov environment from the old East/West environment.  Some of these changes may end up being unwound later on, but for the time being these are the pieces that are required in order to get the app running in the new environment.