Add support to build and manage static assets locally before manual deployments
We have run into issues in the past regarding the automatic collectstatic functionality in the Python buildpack, so we bypass it, but that now requires us to build and manage our static assets locally prior to doing a manual deployment.  Note that we already have Travis setup to do these things as well before it pushes a deployment.