Only show the first paragraph block in the latest update preview
The current behavior is outputting multiple paragraphs if they’re
stored as separate blocks in the body streamfield. Instead of looping
through all body blocks and looking for paragraph block types, this
just checks the first two blocks for a paragraph. This way, if the
first block is an image, it will be ignored. If the first block is a
paragraph, it will just show the one paragraph.