Attempt at squashing migrations
This changeset is an attempt to clean up our migrations and start from scratch with them.  Our app has grown significantly and the state of the database is very different from when we started.  We hope that in doing this it provides a cleaner base to start from now that we have gotten through a lot of heavy feature development.