Fix session timeout issue with cloud.gov
This changeset addresses an issue we have been having with the cloud.gov session/token timeout being reduced to 10 - 15 minutes.  The underlying UAA client we are using, cg-django-uaa, required an update to support refreshing tokens so that sessions remain valid while a user logs in.

A huge thank you to @toolness for providing the fix in cg-django-uaa and helping us address this issue!