Update remaining env var references
This changeset fixes a few more env var references and makes use of the `env.get_credential()` method to ensure that an env var is found.  To recap, that method checks the credential service in Cloud Foundry first for an env var, and if it cannot find it then it falls back to checking the host environment the app is running in.  This method works in both the deployed apps and locally for that reason.