Modified logging code a bit more to account for missing users
I noticed that the instance property had both the user and user_id on it, so the user was already being preloaded if that information was available.  In the cases it is not, I wanted to make sure that the log statement was still of use and reflected what was taking place even if we could not provide the user info, which is the user associated with the PageRevision being processed.

In the case of copy or delete actions, all PageRevision objects are cycled through and processed.  This results in multiple log statements being generated, but these statements maintain the user and date/time associated with the PageRevision.  Whenever a series of statements like this is output, the last one represents the change/action that had just taken place.

There is probably a better way to track these items, but that will require a bit more research and this at least fixes the immediate issue of content managers not being able to perform certain actions like deleting pages.