Improve New Relic initialization
This changeset improves the New Relic initialization to take into account whether or not the New Relic license key is present or not.  In our current setup and workflow, the New Relic license key is always set in the cloud.gov environment, but never locally (unless we are testing this integration explicitly, in which case it is required).  The existing code assumes the license key is always present and attempts to initialize the New Relic Agent, resulting in a multitude of warnings and errors in local development environments that makes it difficult to debug and work.

The setup is copied from our current web app and API New Relic initialization configs.

Lastly, this also adds the pdbpp package for improved command line debugging.