Remove .python-version file
This changeset removes the `.python-version` file from the repository for the time being.  We were never able to move forward on the `pip-tools` install, so this is not needed at the moment and makes dealing with Python updates locally a bit more cumbersome at the moment.  We can bring this back once we have a chance to work on Python dependency management a bit more and provide clear documentation on how to manage these things.

In the interim, if you are using `pyenv` to manage your local Python installations, please be sure you have a global version set, e.g., `pyenv global 3.5.4`.