Fix deploy steps to run commands in proper working directory
This changeset fixes an issue we are having with the deploy step in the CircleCI environment.  It appears that the Django commands are not able to find where certain files are located, particularly those found in the `fec/dist` directory.

There is a larger issue here with the referencing to the path(s) of static assets, but we are fixing the immediate problem first.

Huge h/t to @tbaxter-18f and @cmc333333 for their help in debugging this issue!