Fix folder removal migration
This changeset fixes an issue with migrations that had gotten out of sync.  There was a hotfix done to remove the Folder model that had the correct operations in it, but those got removed in a subsequent merge conflict resolution.  These changes sort all of that out and put things back in the correct order so migrations will run properly again.