Fix static asset paths for rev-manifest.json files
This changeset fixes an issue with static file paths being hardcoded and therefore not being found if things are being run outside of an expected directory.  It leverages the STATIC_ROOT setting in Django to form the first part of the file path referencing the static asset.

The imports have also been tidied up in these two files.