Ensure CSRF cookie is present for feedback tool
This changeset addresses an issue with the feedback tool introduced with the CMS/web app merge.  The cookie with the CSRF token is not being sent or passsed to the view submitting the issue, so this change should make sure that does take place.  See https://docs.djangoproject.com/en/1.10/ref/csrf/#ajax for more details.