Fix for warning banner in data portion of site
This changeset addresses an issue where the environment the application is running in was not being checked for correctly in the data portion of the app since the merge.  This is fixed by creating a new global template tag function that returns the current environment value so that it can be checked in the templates.