Add logging to the API caller
This changeset adds logging to the API caller so we have a bit more visibility into things when an error occurs.  The logging takes place in the underlying `_call_api` method that serves as the backbone for all API requests made by the CMS.

There are several formatting improvements included in this changeset to help improve readability as well.

h/t to @LindsayYoung, who originally did this work!