Updated S3 environment variables to point the bound S3 bucket
This also updates all of the manifests associated with each space to bind the CMS app to the content bucket directly, so we will no longer have to store the credentials in the user credential service.  Once this is merged, each deployment to each environment should update the bindings and take care of this automatically for us.