Added copy for latest updates email signup
The interaction is slightly more complex than we'd realized on the GovDelivery end, so I'm adding a sentence to let folks know there's slightly more.

I don't think the GovDelivery privacy policy is strictly necessary, since a new subscriber will see it on the subsequent page, but I don't think it's bad to keep it anyway. It gives some visual balance to the form while also letting someone see it before doing anything else.

One other thing @johnnyporkchops: can you make sure the text renders well where it is? I'm not sure if putting it before that `ul` is going to mess up the class for that text. We _can_ put it after, though I like it better before.