Solid beta
All data loads are working and syncing. Typeahead, date stamp, state list, state totals, and candidate details are working. Pull-down selector is changing based on selected candidate's election_years. Map is incorrect but it's in and working.
Still serveral TODO items, not the least of which is swapping the map.